feat: implement adjust method focus
